We created a Pub/Sub adapter to avoid having to extend JIT Access with non-core functionality, such as sending messages or logs to third-party systems. This allows defined events to be published to a specific topic.
The events that are currently emitted are:

Binding approval request
Binding implementation, either self or peer approved
Error messages
Events have a unique origin attribute that identifies the event type, as well as a JSON event payload that can be processed by third-party applications. The Pub/Sub functionality is activated when the environment variable TOPIC_NAME is present in the app.yaml file.

This PR also includes an improvement to standardise the time format of the IAM condition to be truncated to seconds, rather than nanos. This allows third-party code in other languages to reliably interpret and process the time format.

Co-authored-by: Mark Statham <mstatham@google.com>
Co-authored-by: Ethan Han <ethanhan@google.com>
Co-authored-by: Ran Zhang <zhangrran@google.com>