b/325372861 Refactor application to use groups
This refactoring forms the basis for JIT Access 2.0, which is
based on groups instead of IAM roles.

The commit introduces the notion of a JIT group. A JIT group is
a Cloud Identity security group that represents a certain job
function (such as Database Administrators for the Foo app)
and bundles any number of IAM roles to one or more resources.

* Membership of a JIT group is always time-bound, but the time
  might vary by group.
* JIT Groups are managed using policies, which use a 3-layer
  hierarchy: Environment > System > Group.
* Optionally, existing "JIT Access 1.x"-style eligible role
  bindings can be mapped to JIT Groups.

The refactoring retains much of the foundation plumbing, but replaces:

* The Core logic to manage and provision access
* The REST API
* The GUI